Remove LockableResource (no use case so far)
