+++ conans/test/unittests/cache/test_cache.py
@@ -215,5 +215,3 @@ def test_concurrent_write_recipe_package(cache: Cache):
     with recipe_layout.lock(blocking=False, wait=True):
         with package_layout.lock(blocking=True, wait=False):
             pass
-
-
